feat: Implement Todo functionality with Todo and TodoItem components
- Added Todo and TodoItem components for task management, allowing users to create, update, and toggle completion of tasks.
- Introduced ITodo and ITodoList types for better type safety and structure.
- Refactored TodoSection to utilize the new Todo component, enhancing modularity and maintainability.
- Improved user interaction with dynamic task addition and display toggling.
