+++ src/types/todo.ts
@@ -1,0 +1,11 @@
+export interface ITodo {
+  id: string;
+  value: string;
+  isCompleted: boolean;
+}
+
+export interface ITodoList {
+  id: string;
+  isShowTodo: boolean;
+  todos: ITodo[];
+}
